finish

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: youskim <youskim@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/24 14:05:54 by youskim           #+#    #+#             */
-/*   Updated: 2022/02/25 13:28:27 by youskim          ###   ########.fr       */
+/*   Updated: 2022/02/25 15:16:52 by youskim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 # include <stdlib.h>
 # include <stdarg.h>
 
-char	*change_hex(unsigned int num);
+char	*change_hex(unsigned long long num);
 int		print_idupx(char type, va_list ap);
 int		ft_printf(const char *format, ...);
 #endif
